+++ frontend-v2/src/App.tsx
@@ -7,6 +7,7 @@ import { DashboardPage } from "./pages/DashboardPage";
 import { FBAShipmentsPage } from "./pages/FBAShipmentsPage";
 import { InventoryPage } from "./pages/InventoryPage";
 import { MasterProductsPage } from "./pages/MasterProductsPage";
+import { MarketplacePage } from "./pages/MarketplacePage";
 import { NotFoundPage } from "./pages/NotFoundPage";
 import { PurchasesPage } from "./pages/PurchasesPage";
 import { SalesPage } from "./pages/SalesPage";
@@ -35,7 +36,7 @@ export function App() {
           <Route path="/fba-shipments" element={<FBAShipmentsPage />} />
           <Route path="/purchases" element={<PurchasesPage />} />
           <Route path="/sales" element={<SalesPage />} />
-          <Route path="/marketplace" element={<StubPage title="Marktplatz" />} />
+          <Route path="/marketplace" element={<MarketplacePage />} />
           <Route path="/cost-allocations" element={<StubPage title="Kostenverteilung" />} />
           <Route path="/opex" element={<StubPage title="Betriebsausgaben" />} />
           <Route path="/mileage" element={<StubPage title="Fahrtenbuch" />} />